【Version_1.4.4】1、修复在修改入错账防错中线程UI及逻辑问题 2、成品出入库的一些工作

+++ app/src/main/java/com/chinashb/www/mobileerp/bean/entity/WCSubProductItemEntity.java
@@ -8,7 +8,7 @@ package com.chinashb.www.mobileerp.bean.entity;
 
 public class WCSubProductItemEntity {
     private WCSubProductEntity wcSubProductEntity;
-    private float qty;
+    private int qty;
     private String buName;
     private String istName;
     private boolean select;
@@ -38,11 +38,11 @@ public class WCSubProductItemEntity {
         this.wcSubProductEntity = wcSubProductEntity;
     }
 
-    public float getQty() {
+    public int getQty() {
         return qty;
     }
 
-    public void setQty(float qty) {
+    public void setQty(int qty) {
         this.qty = qty;
     }
